Fixes display of widget on pages and in posts.
The problem was that the blip was being echoed by the function that constructs the blip (blipper_widget_display_blip). The fix is to have a function (get_blipper_widget_display_blip) construct the blip into a string () and return it completed blip. The original function (blipper_widget_display_blip) still needs to echo the blip, so that echos the output of the new function (get_blipper_widget_display_blip). The new function is essentially the old function, but renamed and changed so that it returns a variable instead of echoing everything.