Updates and improves the readme file.
Moves the lesser changes from the change log to a new change log file.

Refers to WordPress' plugin installation instructions for all methods of installation in one go.

Adds instructions for the shortcode: refers to WordPress for how to use shortcodes in general, and gives examples of how to use [blipper_widget].

Makes other changes and improvements to the wording, removing much redundancy.