Allows the user to optionally display the blip's text on the shortcode version.
Improves the styling: allows the user to use the widget form settings to style the widget or to use CSS.  Removes the widget styling from the shortcode, which is now styled only with CSS.  Adds CSS classes to the relevant elements for easier styling.

Fixes the styling of the widget (the hard-coded inline styles).

Fixes the HTML output according to Nu Html Checker by W3C.