Updates the version number to 1.2.6-alpha.
Version updated because it's not v1.2.5 any more, but is still tagged as 1.2.5 in git because of the way it works wrt releases, though.

Changes the plugin data method from private to public so it can be used by the widget class. Allows the data to be obtained whether the plugin has been loaded or not. Moves the version constant to inside the settings class and adds a plugin-name constant.