Changes rendering source output to BW log from error log.
That is, the debug output that says whether the blip is rendered from scratch or from the cache is now output by the BW log function rather than PHP's error log function. This is so that it isn't accidently printed when BW_DEBUG is false: this debug output is commonly displayed.