Deletes the BW transients when widget settings updated.
When the widget settings are changed, a new transient is created to cache the blip. This can result in loads of BW transients clutterint the place up. It would be best to delete only theone just created, but this can't be done just yet, so all of them are deleted for now.

Contributes to BW-47.