Changes the way the cache key works.
The cache key is no longer stored statically because it doesn't work. It can't be stored non-statically because static functions need it. It must therefore be generated afresh each time one is needed.

The way the cache key for caching blips is generated has also been changed: it now takes into account only the settings array. The expiration time and any blip title have been removed.

There's now a method to display the status of the cache as debug output if desired.

Contributes to BW-47 and BW-53.